Fix Distress Beacon Automatic Shuttle Handling (#3196)
# About the pull request

This PR changes the default shuttle_id and fixes an error in the launch
checks for distress beacons. Many distress signals such as the xeno erts
do not override the value but need an automatic launch procedure. This
also tweaks the logging sent to admins in case of an automatic launch
failure.

# Explain why it's good for the game

Fixes #2283 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![ert](https://user-images.githubusercontent.com/76988376/235391585-18648827-3e38-472d-a1d7-ea3de1f42ab7.gif)

</details>


# Changelog
:cl: Drathek
fix: Fixes automatic shuttle launching for distress signals that require
it (Xeno & Zombie ERTs)
/:cl: