Allows for surgery in yautja armour (#3242)
# About the pull request
Allows preds to use their medicomp surgery kit while their armour is on
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The new changes to surgery have really shoved a stick into the spokes of
pred gameplay in what is probably an unintended way as if the pred has a
back piece such as a cape/poncho then their weapon will have to go to
the floor, saying there is equipment to be collected for other preds and
distracting them

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
balance: Allows preds to use medicomp surgery with their armour on
/:cl: