Fixed the weapon examine strength message being wrong in some cases (#3246)
# About the pull request

When checking the throw force the default message was never set back to
"weak", so it would end up showing throw force the same strength as the
melee force if throwing force was at least 5 but lower than 10.

Also made the checks more consistent and not overlap.

# Explain why it's good for the game

No.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed an edge case that caused the melee strength's value to be
shown as the throwing strength, when examining a weapon.
/:cl:

---------

Co-authored-by: Rigó János <rigojanos@protonmail.com>