Brings back the old hidden riot suit  (#3243)
# About the pull request
The recent change of mp clothes changed the old hidden riot suit that is
on the almayer. I decided to bring it back so that it is still different
from the mp's one while keeping the charm it always had. It doesn't
affect the mp's wardrobe at all and is solely targeting the Almayer
"easter egg" one .
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
There are points that can be made about it being backward but I really
liked the old sprite and it's a cool maintenance shaft loot. Also stops
regular marines from larping as the military police.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Molto
add: Re-added the old riot armor on the Almayer
/:cl: