Fix re-zooming when observing a mark or xeno (#3252)
# About the pull request

This PR prevents using the toggle_long_range ability when observing a
xeno or mark. There is already behavior that causes the zoom effect to
end on observation, but nothing was preventing zooming again while still
in that state.

Entering a tunnel, and traveling to other locations in the tunnel does
not reset any zoom, nor prevents zooming. I think this is fine because
you are actually in that location and are using your own senses to view
further.

# Explain why it's good for the game

Fixes #356 

# Changelog
:cl: Drathek
fix: Fixed xeno toggle_long_range ability from being used when observing
a xeno or mark
/:cl: