Set ATTACHED to true after the hugger is dropped
So that manually infecting targets by clicking on them with the hugger in hand no longer causes the huggers' immediate death and permanent stay on the victim's face.

The comment on the "equipped" proc also now properly reflects what calling the "go_idle()" proc actually does to the hugger when equipped.