Remove Dead Code for Lurker Rush (#3300)
# About the pull request

This PR simply removes a line in lurker rush code that is currently
doing nothing (it just hits the return statement because a timer for the
cooldown is already active). See #2938 on how it should have been, but
it currently fits fine in balance. If it were to be a thing, this
ability would likely need to only work *at all* on a valid sprite click.
Wiki page has already been updated to remove mention that a successful
attack reduces the cooldown.

# Explain why it's good for the game

Code should convey what it is doing, not suggest it is doing something
else.

# Changelog
:cl: Drathek
code: Removed dead code in lurker rush suggesting it lowered cooldown on
a successful hit
/:cl: