fix Trijent elevator will now need to have power to work. (#3329)
# About the pull request

Fixing elevator working even if the area as is APC cut.
Fix buttons to call elevator not needing to fix power locally to use
them.

fixes https://github.com/cmss13-devs/cmss13/issues/3312

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

The elevator is more consistent with the rest of the map it require
power to function.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:

fix: fix being able to launch elevator without fixing power in the
concerned areas.
code: connected the APC inside the elevator to the power net ground.
code: add the ability for the elevator area to loose power.
maptweak: move all the call buttons for the elevator inside area that
have APC.

/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>