Maps for squaddies (#3334)
# About the pull request

This PR adds handheld maps to all squad vendors.

# Explain why it's good for the game

Easy to access in game references are important. A wiki should not be a
pseudo requirement to play or learn the game effectively.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Handheld maps added to all squad marine vendors.
/:cl:

Co-authored-by: harryob <me@harryob.live>