Renames Solaris Ridge walls (#3338)
# About the pull request
Currently all the walls on solaris ridge arbitrarily reference the
colony name, and don't indicate if they're reinforced or not. This
changes that.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Consistency, and makes the walls more re-usable without having to
instance their names.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
spellcheck: Adjusted names of Solaris Ridge walls to represent what they
are without using the colony's name.
/:cl: