Move the DATAbase to data (#3351)
# About the pull request

This PR simply moves the local database file location from root into the
data folder.

# Explain why it's good for the game

This came up as an issue for me because there's some weird issues with
checking out old versions of the code vs the present which corrupts the
database. Clearing out the data folder was insufficient to fix the issue
because the database exists in the root some reason, so its just
confusing that its there and not the data folder.

# Changelog
:cl: Drathek harryob
server: Moved the local.db database location from root to the data
folder
/:cl: