Adds a warning to cryomarines about ghosting (#3353)
# About the pull request
People keep spawning in as foxtrot and cryoing/ghosting if they don't
get a free spec or special role. This can greatly impact the use of the
ERT. If players ahelp as they're meant to, we can replace them.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Stops wasting the cryomarines.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a warning/reminder to cryomarine spawns to ahelp if they want
to cryo/ghost upon spawning in, to stop people wasting the ERT call.
/:cl: