Fix bug that didn't remove luminosity after the object that emit them was destroy. (#3381)
# About the pull request

fixes https://github.com/cmss13-devs/cmss13/issues/3371

added a check if the entity is destroy that reset luminosity to 0.
adding to the code their own entity instead of doing change in dmm file
for :
the cart and the bio luminescent tree also did it for glow goo that give
acid damage.
give the proper entity name instead of doing change in dmm files for the
light stick

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
remove a bug.
less useless code in dmm files...

# Testing Photographs and Procedure
i tested on local server it work very fine.
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:

fix: remove the luminosity when torch and plant that produce the light
are destroy.

/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>