Fix the dead will now be properly turn into zombie when infected by black go virus. (#3340)
# About the pull request

code: changed longevity of black go to 500 instead of 150 so that it as
time to turn the dead into zombie.

fixes https://github.com/cmss13-devs/cmss13/issues/3331

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

fixing bug is good.
# Testing Photographs and Procedure

i tested on a local server it worked fine.
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:

fix: fix the dead will now be properly turn into zombie when infected by
black go virus.

/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>