Fixes for boiler. (#3390)
# About the pull request

If zooming with the action applied some effects, unzooming should cancel
them out, whether the unzooming is caused by using the action again or
by something else.
Trapper inheriting some things added to base boiler in globber update is
simply an oversight. Its vision has always been centered on the trapper
and should be, as it is lined exactly with its bombardment distance (and
currently trying to use acid mine at the edge of offset range, which is
outside of intended zoom range, gives a warning that "something is in
the way" instead of saying that it is out of range, which makes it even
worse for an unaware player).
Likewise, adding shroud for the globber should have been restricted to
the globber - instead, it persists on trapper and hijacks its fourth
keybind which is meant to be its zooming ability.

# Explain why it's good for the game

Fixes #3385.


# Changelog
:cl:
fix: Boiler's slowdown for zooming now properly resets on unzooming
regardless of what causes it.
fix: Trapper's zoomed vision is no longer unintentionally offset.
fix: Trapper no longer gets globber's Acid Shroud ability messing its
keybinds.
/:cl: