Yautja armor rebalance (#3303)
# About the pull request


https://docs.google.com/spreadsheets/d/1ywbo-AhXE3bALFb4y0VtaiXWKIlCgbhTIi7mZlvRVsk/edit?usp=sharing

Will update CL before merge as this is going to change a bit during
testing likely. In the meantime, the sheets page linked will be updated.

As a reminder, design/balance feedback goes here:
https://forum.cm-ss13.com/w/pr-feedback/steps/step_1

As a second reminder: Unless I've reached out to you to discuss this PR
please don't ping me. I will read through feedback given as I can. I get
a metric ton of messages a day.

# Explain why it's good for the game

The extreme variance between people who are aware of how predators
function versus people who are unaware of how predators function when it
comes to fighting is not enjoyable for the majority of the playerbase.
Knowledge gap should be important, it should not be the difference
between doing good damage and quite literally zero damage. Hopefully by
the end of this we will have a predator that requires care and
consistency to play rather than a walking tank for new people to plink
at while avoiding all the good players.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Yautja armor rebalance
/:cl: