Fix Shuttle Landing Lights and Looming Shadow (#3417)
# About the pull request

This PR restores the hyperspace ripple effect for shuttles that are
landing, as well as their landing lights getting stuck on (though will
be more obvious when #3305 is merged because it fixed its lingering
references to landing lights when it gets crushed by the shuttle). A
debug message should also be logged whenever an anchored thing is
crushed by the shuttle: There's a few maps where they have stairs and
such that either should get whitelisted here too, or just the map
adjusted to have a properly cleared landing zone.

# Explain why it's good for the game

- Better indication where its not safe to be (else get gibbed)
- It looks fancy again (see below section)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



![dark](https://github.com/cmss13-devs/cmss13/assets/76988376/2db4c846-52c4-4099-9857-b5d12f8ac8e9)


</details>


# Changelog
:cl: Drathek
fix: Fix shuttle landing lights getting stuck on
imageadd: Added a transparent gray sprite to act as a shadow when a
shuttle or something is looming over that tile
/:cl: