increases the synthetic piloting skill (#3419)
# About the pull request

Fixes the oversight when we changed to TGUI pilot interface which made
it so pilot level 2 is required for flybys and optimized flight.
Confirmed with maintaining management that it was an oversight.

# Explain why it's good for the game

Lets synthetics perform duties like medevac and fulton without it being
jank, by allowing flybys.


# Testing Photographs and Procedure

</details>


# Changelog

:cl:
balance: synthetic pilot skill increased from 1 to 2
/:cl: