Corpsmen buyable blood bags (#3459)
# About the pull request

This PR allows corpsmen to buy blood bags in their vendors.

# Explain why it's good for the game

Blood bags are becoming increasingly more required in our medical system
and more are going to be needed.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Adds buyable blood bags to corpsmen vendors
/:cl: