Revival of #2979: Ammo Can IMP Rack (#3392)
# About the pull request

Adds a relatively scarce logistics backpack item to the req vendor,
allows Marines to carry 3 of any obj/item/ammo_box/, increases the
logistics capability of your standard Marine. (Special thanks to
Sionnach for the sprites, and Bengus, Harry, and Beagle for help with
the code!)

# Explain why it's good for the game

Instead of 400 boxes of ammo or mags remaining at the FOB, this new item
encourages logistics and ammo-bearing (A lot of riflemen like wearing
ammo boxes on their backs, this allows them to hold more boxes + MRE
boxes/flare boxes as well, thanks to those being a subtype of
/ammo_box/. Allows the movement of supplies easier if you can get your
hands on this elusive, limited issue item!

# Testing Photographs and Procedure


![image](https://github.com/cmss13-devs/cmss13/assets/112824777/56d3ab1f-daec-4cf3-ac66-e6f08efe0162)

![image](https://github.com/cmss13-devs/cmss13/assets/112824777/59693ccc-6ad3-40f6-ae0a-f4a1afdfab01)

![image](https://github.com/cmss13-devs/cmss13/assets/112824777/25e67432-a63c-41a0-a53c-0a4183d17fe6)

# Changelog

:cl:LynxSolstice, Sionnach
add: Adds a new IMP framed back item, the Ammo Rack.
balance: Added the new Ammo Rack to the Req vendor. At a scaling of
(round*1), it should be pretty scarce. The ammo rack also is unthrowable
and adds a flat .5 second increase to movement delay.
imageadd: Added icons for Ammo_Pack_ and onmobs for each state (0, 1, 2,
and 3)
/:cl:

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>
Co-authored-by: fira <loyauflorian@gmail.com>