Fix trijent shuttle apc can now be fixed. (#3462)
# About the pull request

since you can't crowbar floor on shuttles you can't see the cable link
to apc and cut them to make a new APC...
so i added a scenery to explain why this spot lack a tile like around
it...

fixes: https://github.com/cmss13-devs/cmss13/issues/3461

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
because i added a small scene and it fix a bug
# Testing Photographs and Procedure
<details>


![image](https://github.com/cmss13-devs/cmss13/assets/117036822/4043a495-e100-4f65-94ba-994af3bc48be)

<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: trijent shuttle apc can now be properly fixed...
maptweak: added a scenery around the fix to make it look more natural.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>