Canceling for the hive now resets its slowdown (#3449)
# About the pull request

This PR clears the slowdown effect applied when for the hive is
canceled.

# Explain why it's good for the game

Though we probably want it to be punishing canceling a for the hive, it
still costs acid so it can feel bugged if you remain slowed for 26-40
seconds depending on when you cancel the ability. Closes #3446

# Changelog
:cl: Drathek
balance: Acid Runner's For the Hive ability now clears its slowdown
effect when canceled
/:cl: