LV522 OB Changes (#3473)
# About the pull request

This PR allows every area but the cylindrical hive areas to be OB'd.

# Explain why it's good for the game

Due to changes to the map you have to dive deep to some hive locations
and it makes it incredibly difficult without effective OBs.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Made more areas in LV522 reactor OB-able
/:cl: