Fixes the Solaris PMC nightmare's issue with locked rifles. (#3343)
# About the pull request

Fixes #3325 , adds the new rifle/m41a/corporate/no_lock variant as a
spawnable preset for PMC survivors.

# Explain why it's good for the game

Fixes a bug and makes sure the fix isn't inconsistent with the gear the
PMCs have available.

# Testing Photographs and Procedure
Runs and compiles fine.

# Changelog

:cl: LynxSolstice
maptweak: Replaced the IFF locked M41As in the PMC nightmare that spawn
on the ground with the new no_lock variant.
/:cl:

---------

Co-authored-by: morrowwolf <darthbane97@gmail.com>