Nerfs lurker pounce movement and adds a cooldown for headbite. (#3496)
# About the pull request
I removed the fact that slashing as lurker allows you to move
immediately after pouncing someone while in cloak, and headbite now has
a 10 second CD. Both of these things have balance approval from Morrow.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The free movement thing was kind of ridicilous and headbite is an
ability that seriously hurts whoever you use it against, so having a CD
makes sense.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Headbite has a 10 second cooldown.
/:cl: